// frontend/src/app/generated/api/model/parsedCvData.ts
/**
 * Yaam.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ParsedWorkExperienceData } from './parsedWorkExperienceData';
import { ParsedEducationData } from './parsedEducationData';
import { ParsedInfoData } from './parsedInfoData';
import { ParsedCertificationData } from './parsedCertificationData';
import { ParsedLanguageData } from './parsedLanguageData';

export interface ParsedCvData {
  info: ParsedInfoData | null;
  workExperiences: Array<ParsedWorkExperienceData>;
  educations: Array<ParsedEducationData>;
  skills: Array<string>;
  languages: Array<ParsedLanguageData>;
  certifications: Array<ParsedCertificationData>;
}